Document schema type analysis: 96% perfect match with Apify
Created comprehensive analysis of database schema vs Apify output types.

Findings:
- 50/52 fields have exact type matches ✅
- 2/52 fields have minor mismatches (INTEGER vs BIGINT) ⚠️
  * ai_work_arrangement_office_days (stores 0-7)
  * ai_working_hours (stores 0-80)
- No functional impact - INTEGER handles these ranges perfectly
- No migration needed - would require dropping views for zero benefit

Updated DATABASE_SCHEMA.sql to document correct types (INTEGER→BIGINT)
for reference, though migration not performed due to view dependencies.

Created docs/SCHEMA_TYPE_ANALYSIS.md with complete field mapping table
showing all 52 Apify fields and their PostgreSQL type equivalents.

Conclusion: Schema is production-ready with 96% exact match rate.

+++ DATABASE_SCHEMA.sql
@@ -78,7 +78,7 @@ CREATE TABLE jobmarket_jobs (
   -- AI Job Classification
   ai_experience_level TEXT,                       -- 0-2, 2-5, 5-10, 10+
   ai_work_arrangement TEXT,                       -- Remote Solely/Remote OK/Hybrid/On-site
-  ai_work_arrangement_office_days INTEGER,
+  ai_work_arrangement_office_days BIGINT,
   ai_remote_location TEXT[],
   ai_remote_location_derived TEXT[],
   
@@ -91,7 +91,7 @@ CREATE TABLE jobmarket_jobs (
   -- AI Job Details
   ai_core_responsibilities TEXT,                  -- 2-sentence summary
   ai_requirements_summary TEXT,                   -- 2-sentence summary
-  ai_working_hours INTEGER DEFAULT 40,
+  ai_working_hours BIGINT DEFAULT 40,
   ai_employment_type TEXT[],
   ai_job_language TEXT,
   ai_visa_sponsorship BOOLEAN,
